feat(cli): upgrade to @angular@cli@6
