+++ src/app/components/components/markdown/markdown.component.html
@@ -85,12 +85,12 @@
           [teradata github Landing]: https://teradata.github.io/
 
           Inline image
-          ![alt text here](app/assets/icons/teradata.svg)
+          ![alt text here](assets/icons/teradata.svg)
 
           Reference-style image:
           ![alt text][logo]
 
-          [logo]: app/assets/icons/teradata.svg "Teradata Labs"
+          [logo]: assets/icons/teradata.svg "Teradata Labs"
         </td-markdown>
       </mat-tab>
       <mat-tab>
@@ -105,12 +105,12 @@
               [teradata github Landing]: https://teradata.github.io/
 
               Inline image
-              ![alt text here](app/assets/icons/teradata.svg)
+              ![alt text here](assets/icons/teradata.svg)
 
               Reference-style image:
               ![alt text][logo]
 
-              [logo]: app/assets/icons/teradata.svg "Teradata Labs"
+              [logo]: assets/icons/teradata.svg "Teradata Labs"
             </td-markdown>
           ]]>
         </td-highlight>
