+++ src/app/components/docs/icons/icons.component.html
@@ -23,7 +23,7 @@
        constructor(private _iconRegistry: MatIconRegistry,
                    private _domSanitizer: DomSanitizer) {
          this._iconRegistry.addSvgIconInNamespace('assets', 'sun',
-           this._domSanitizer.bypassSecurityTrustResourceUrl('app/assets/icons/sun.svg'));
+           this._domSanitizer.bypassSecurityTrustResourceUrl('assets/icons/sun.svg'));
        }
      ]]>
     </td-highlight>
