+++ src/app/components/layouts/overview/overview.component.html
@@ -7,7 +7,7 @@
     <div layout="row" layout-padding>
       <ng-template let-item let-last="last" ngFor [ngForOf]="items">
         <a mat-button flex [routerLink]="[item.route]">
-          <img src="app/assets/images/{{item.route}}.png" alt="{{item.title}}">
+          <img src="assets/images/{{item.route}}.png" alt="{{item.title}}">
           <div class="mat-subtitle">{{item.title}}</div>
           <div class="mat-body-1 text-wrap text-left">{{item.description}}</div>
         </a>
