+++ src/app/components/style-guide/resources/resources.component.html
@@ -11,19 +11,19 @@
   <mat-divider></mat-divider>
   <mat-nav-list class="pad-top-none">
     <a mat-list-item target="_blank" href="https://github.com/Teradata/covalent-resources/raw/develop/axure/covalent-axure.rplib">
-      <img matListIcon src="app/assets/icons/axure.png" alt="Axure icon">
+      <img matListIcon src="assets/icons/axure.png" alt="Axure icon">
       <h3 matLine>Axure RP Libraries</h3>
       <p matLine>Covalent libraries for the Axure RP</p>
     </a>
     <mat-divider></mat-divider>
     <a mat-list-item target="_blank" href="https://github.com/Teradata/covalent-resources/raw/develop/illustrator/covalent-illustrator.ai">
-      <img matListIcon src="app/assets/icons/illustrator.svg" alt="Illustrator icon">
+      <img matListIcon src="assets/icons/illustrator.svg" alt="Illustrator icon">
       <h3 matLine>Illustrator Template</h3>
       <p matLine>Covalent template for the Adobe Illustrator</p>
     </a>
     <mat-divider></mat-divider>
     <a mat-list-item target="_blank" href="https://github.com/Teradata/covalent-resources/raw/develop/sketch/covalent-sketch.sketch">
-      <img matListIcon src="app/assets/icons/sketchapp.svg" alt="Sketch icon">
+      <img matListIcon src="assets/icons/sketchapp.svg" alt="Sketch icon">
       <h3 matLine>Sketch Template</h3>
       <p matLine>Covalent template for the macOS Sketch App</p>
     </a>
