+++ src/assets/i18n/es.json
