+++ src/index.html
@@ -8,11 +8,11 @@
   <link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet">
   <meta name="viewport" content="width=device-width, initial-scale=1">
 
-  </head>
+</head>
 <body>
   <docs-covalent>
     <div style="padding: 20%;text-align:center;height:100%;background-color:#546E7A;color:#B0BEC5;">
-      <img alt="Covalent Logo" src="app/assets/icons/covalent.svg" width="100">
+      <img alt="Covalent Logo" src="assets/icons/covalent.svg" width="100">
       <h3>Covalent Loading...</h3>
     </div>
   </docs-covalent>
