+++ src/tsconfig.spec.json
@@ -9,8 +9,12 @@
       "jasmine", "hammerjs", "node"
     ]
   },
-  "include": [
+  "files": [
     "test.ts",
-    "**/*.spec.*"
+    "polyfills.ts"
+  ],
+  "include": [
+    "**/*.spec.ts",
+    "**/*.d.ts"
   ]
 }
