--- src/typings.d.ts
@@ -1,6 +1,0 @@
-// Typings reference file, see links for more information
-// https://github.com/typings/typings
-// https://www.typescriptlang.org/docs/handbook/writing-declaration-files.html
-
-declare var System: any;
-declare var require: any;
